+++ src/main/java/com/example/book_your_seat/coupon/facade/CouponQueryFacade.java
@@ -2,7 +2,7 @@ package com.example.book_your_seat.coupon.facade;
 
 import com.example.book_your_seat.coupon.controller.dto.CouponDetailResponse;
 import com.example.book_your_seat.coupon.domain.UserCoupon;
-import com.example.book_your_seat.coupon.manager.UserManager;
+import com.example.book_your_seat.user.manager.UserManager;
 import com.example.book_your_seat.user.domain.User;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
